pyproject.toml
[build-system]
requires = ["flit_core>=3.9,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "django-ninja-plus"
dynamic = ["version"]
description = "A fork of django-ninja-extra for Django Ninja with class-based controllers and advanced API patterns."
readme = "README.md"
requires-python = ">=3.14,<3.15"
license = { file = "LICENSE" }
authors = [
    { name = "Diego Saraiva", email = "diegosaraiva@gmail.com" },
]
keywords = ["django", "django-ninja", "api", "rest", "controllers"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Framework :: AsyncIO",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Internet",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "Django>=5.0",
    "django-ninja>=1.6.0",
    "injector>=0.19.0",
    "asgiref>=3.8.1",
]

[project.urls]
Documentation = "https://github.com/jdiego/django-ninja-plus"
Repository = "https://github.com/jdiego/django-ninja-plus"
Homepage = "https://github.com/jdiego/django-ninja-plus"

[tool.flit.module]
name = "ninja_plus"

[dependency-groups]
test = [
    "django-stubs",
    "mypy==1.14.1",
    "ninja-schema>=0.14.1",
    "pytest>=8.0",
    "pytest-asyncio==1.3.0",
    "pytest-cov",
    "pytest-django",
    "ruff==0.15.4",
]
docs = [
    "markdown-include",
    "mdx-include>=1.4.1,<2.0.0",
    "mkdocs>=1.1.2,<2.0.0",
    "mkdocs-markdownextradata-plugin>=0.1.7,<0.3.0",
    "mkdocs-material>=7.1.9,<10.0.0",
    "mkdocstrings",
]
dev = [
    { include-group = "test" },
    { include-group = "docs" },
    "pre-commit",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "C",
    "B",
]
ignore = [
    "E501",
    "B008",
    "C901",
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.isort]
known-first-party = ["ninja_plus", "ninja_extra"]
known-third-party = ["django", "ninja", "injector", "pydantic", "asgiref"]

[tool.mypy]
python_version = "3.14"
show_column_numbers = true
follow_imports = "normal"
ignore_missing_imports = true
disallow_untyped_calls = true
warn_return_any = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unused_ignores = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_reexport = true
disable_error_code = ["var-annotated", "misc"]

[[tool.mypy.overrides]]
module = "ninja_extra.compatible.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "ninja_plus.compatible.*"
ignore_errors = true


